remove BQJobName, move JobLabels into own file
