+++ core/src/main/scala/no/nrk/bigquery/metrics/MetricsOps.scala
@@ -8,45 +8,45 @@ package no.nrk.bigquery.metrics
 
 import cats.Applicative
 import com.google.cloud.bigquery.JobStatistics
-import no.nrk.bigquery.BQJobName
+import no.nrk.bigquery.BQJobId
 
 trait MetricsOps[F[_]] {
-  def increaseActiveJobs(jobName: BQJobName): F[Unit]
-  def decreaseActiveJobs(jobName: BQJobName): F[Unit]
-  def recordTotalTime(elapsed: Long, jobName: BQJobName): F[Unit]
+  def increaseActiveJobs(jobName: BQJobId): F[Unit]
+  def decreaseActiveJobs(jobName: BQJobId): F[Unit]
+  def recordTotalTime(elapsed: Long, jobName: BQJobId): F[Unit]
   def recordAbnormalTermination(
       elapsed: Long,
       terminationType: TerminationType,
-      jobName: BQJobName
+      jobName: BQJobId
   ): F[Unit]
   def recordTotalBytesBilled(
       job: Option[JobStatistics],
-      jobName: BQJobName
+      jobName: BQJobId
   ): F[Unit]
 }
 
 object MetricsOps {
   def noop[F[_]](implicit F: Applicative[F]): MetricsOps[F] = new MetricsOps[F] {
-    override def increaseActiveJobs(jobName: BQJobName): F[Unit] =
+    override def increaseActiveJobs(jobName: BQJobId): F[Unit] =
       F.unit
 
-    override def decreaseActiveJobs(jobName: BQJobName): F[Unit] =
+    override def decreaseActiveJobs(jobName: BQJobId): F[Unit] =
       F.unit
 
     override def recordTotalTime(
         elapsed: Long,
-        jobName: BQJobName
+        jobName: BQJobId
     ): F[Unit] = F.unit
 
     override def recordAbnormalTermination(
         elapsed: Long,
         terminationType: TerminationType,
-        jobName: BQJobName
+        jobName: BQJobId
     ): F[Unit] = F.unit
 
     override def recordTotalBytesBilled(
         job: Option[JobStatistics],
-        jobName: BQJobName
+        jobName: BQJobId
     ): F[Unit] = F.unit
   }
 }
