+++ core/src/main/scala/no/nrk/bigquery/syntax.scala
@@ -6,6 +6,6 @@
 
 package no.nrk.bigquery
 
-import no.nrk.bigquery.internal.{BQLiteralSyntax, BQNameConversions, BQShowSyntax}
+import no.nrk.bigquery.internal.{BQLiteralSyntax, BQShowSyntax}
 
-object syntax extends BQLiteralSyntax with BQShowSyntax with BQNameConversions
+object syntax extends BQLiteralSyntax with BQShowSyntax
