+++ testing/src/main/scala/no/nrk/bigquery/testing/BQUdfSmokeTest.scala
@@ -78,7 +78,7 @@ object BQUdfSmokeTest {
           )
 
           val run = bqClient
-            .synchronousQuery(BQJobId(None, None, BQJobName("smoketest"), JobLabels.empty), BQQuery[Json](query))
+            .synchronousQuery(BQJobId("smoketest"), BQQuery[Json](query))
             .compile
             .lastOrError
             .guaranteeCase {
